Network.AWS.Inspector.CreateAssessmentTarget

Description

Creates a new assessment target using the ARN of the resource group that is generated by CreateResourceGroup . You can create up to 50 assessment targets per AWS account. You can run up to 500 concurrent agents per AWS account. For more information, see Amazon Inspector Assessment Targets .

Creating a Request

createAssessmentTarget

Arguments

:: Text
catAssessmentTargetName

-> Text
catResourceGroupARN

-> CreateAssessmentTarget

Creates a value of CreateAssessmentTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

catAssessmentTargetName - The user-defined name that identifies the assessment target that you want to create. The name must be unique within the AWS account.

catResourceGroupARN - The ARN that specifies the resource group that is used to create the assessment target.

data CreateAssessmentTarget

See: createAssessmentTarget smart constructor.

Request Lenses

catAssessmentTargetName :: Lens' CreateAssessmentTarget Text

The user-defined name that identifies the assessment target that you want to create. The name must be unique within the AWS account.

catResourceGroupARN :: Lens' CreateAssessmentTarget Text

The ARN that specifies the resource group that is used to create the assessment target.

Destructuring the Response

createAssessmentTargetResponse

Arguments

:: Int
catrsResponseStatus

-> Text
catrsAssessmentTargetARN

-> CreateAssessmentTargetResponse

Creates a value of CreateAssessmentTargetResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

catrsResponseStatus - -- | The response status code.

catrsAssessmentTargetARN - The ARN that specifies the assessment target that is created.

data CreateAssessmentTargetResponse

See: createAssessmentTargetResponse smart constructor.

Response Lenses

catrsResponseStatus :: Lens' CreateAssessmentTargetResponse Int

- | The response status code.

catrsAssessmentTargetARN :: Lens' CreateAssessmentTargetResponse Text

The ARN that specifies the assessment target that is created.