Network.AWS.Inspector.CreateAssessmentTemplate

Description

Creates an assessment template for the assessment target that is specified by the ARN of the assessment target.

Creating a Request

createAssessmentTemplate

Arguments

:: Text
catAssessmentTargetARN

-> Text
catAssessmentTemplateName

-> Natural
catDurationInSeconds

-> CreateAssessmentTemplate

Creates a value of CreateAssessmentTemplate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

catUserAttributesForFindings - The user-defined attributes that are assigned to every finding that is generated by the assessment run that uses this assessment template. An attribute is a key and value pair (an Attribute object). Within an assessment template, each key must be unique.

catAssessmentTargetARN - The ARN that specifies the assessment target for which you want to create the assessment template.

catAssessmentTemplateName - The user-defined name that identifies the assessment template that you want to create. You can create several assessment templates for an assessment target. The names of the assessment templates that correspond to a particular assessment target must be unique.

catDurationInSeconds - The duration of the assessment run in seconds. The default value is 3600 seconds (one hour).

catRulesPackageARNs - The ARNs that specify the rules packages that you want to attach to the assessment template.

data CreateAssessmentTemplate

See: createAssessmentTemplate smart constructor.

Request Lenses

catUserAttributesForFindings :: Lens' CreateAssessmentTemplate [Attribute]

The user-defined attributes that are assigned to every finding that is generated by the assessment run that uses this assessment template. An attribute is a key and value pair (an Attribute object). Within an assessment template, each key must be unique.

catAssessmentTargetARN :: Lens' CreateAssessmentTemplate Text

The ARN that specifies the assessment target for which you want to create the assessment template.

catAssessmentTemplateName :: Lens' CreateAssessmentTemplate Text

The user-defined name that identifies the assessment template that you want to create. You can create several assessment templates for an assessment target. The names of the assessment templates that correspond to a particular assessment target must be unique.

catDurationInSeconds :: Lens' CreateAssessmentTemplate Natural

The duration of the assessment run in seconds. The default value is 3600 seconds (one hour).

catRulesPackageARNs :: Lens' CreateAssessmentTemplate [Text]

The ARNs that specify the rules packages that you want to attach to the assessment template.

Destructuring the Response

createAssessmentTemplateResponse

Arguments

:: Int
crsResponseStatus

-> Text
crsAssessmentTemplateARN

-> CreateAssessmentTemplateResponse

Creates a value of CreateAssessmentTemplateResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsResponseStatus - -- | The response status code.

crsAssessmentTemplateARN - The ARN that specifies the assessment template that is created.

data CreateAssessmentTemplateResponse

See: createAssessmentTemplateResponse smart constructor.

Response Lenses

crsResponseStatus :: Lens' CreateAssessmentTemplateResponse Int

- | The response status code.

crsAssessmentTemplateARN :: Lens' CreateAssessmentTemplateResponse Text

The ARN that specifies the assessment template that is created.