Network.AWS.Inspector.CreateResourceGroup

Description

Creates a resource group using the specified set of tags (key and value pairs) that are used to select the EC2 instances to be included in an Amazon Inspector assessment target. The created resource group is then used to create an Amazon Inspector assessment target. For more information, see CreateAssessmentTarget .

Creating a Request

createResourceGroup

Arguments

:: NonEmpty ResourceGroupTag
crgResourceGroupTags

-> CreateResourceGroup

Creates a value of CreateResourceGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crgResourceGroupTags - A collection of keys and an array of possible values, '[{"key":"key1","values":[Value1,Value2]},{"key":Key2,"values":[Value3]}]'. For example,'[{"key":Name,"values":[TestEC2Instance]}]'.

data CreateResourceGroup

See: createResourceGroup smart constructor.

Request Lenses

crgResourceGroupTags :: Lens' CreateResourceGroup (NonEmpty ResourceGroupTag)

A collection of keys and an array of possible values, '[{"key":"key1","values":[Value1,Value2]},{"key":Key2,"values":[Value3]}]'. For example,'[{"key":Name,"values":[TestEC2Instance]}]'.

Destructuring the Response

createResourceGroupResponse

Arguments

:: Int
crgrsResponseStatus

-> Text
crgrsResourceGroupARN

-> CreateResourceGroupResponse

Creates a value of CreateResourceGroupResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crgrsResponseStatus - -- | The response status code.

crgrsResourceGroupARN - The ARN that specifies the resource group that is created.

data CreateResourceGroupResponse

See: createResourceGroupResponse smart constructor.

Response Lenses

crgrsResponseStatus :: Lens' CreateResourceGroupResponse Int

- | The response status code.

crgrsResourceGroupARN :: Lens' CreateResourceGroupResponse Text

The ARN that specifies the resource group that is created.