Network.AWS.Inspector.DeleteAssessmentTarget

Description

Deletes the assessment target that is specified by the ARN of the assessment target.

Creating a Request

deleteAssessmentTarget

Arguments

:: Text
datAssessmentTargetARN

-> DeleteAssessmentTarget

Creates a value of DeleteAssessmentTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

datAssessmentTargetARN - The ARN that specifies the assessment target that you want to delete.

data DeleteAssessmentTarget

See: deleteAssessmentTarget smart constructor.

Request Lenses

datAssessmentTargetARN :: Lens' DeleteAssessmentTarget Text

The ARN that specifies the assessment target that you want to delete.

Destructuring the Response

deleteAssessmentTargetResponse :: DeleteAssessmentTargetResponse

Creates a value of DeleteAssessmentTargetResponse with the minimum fields required to make a request.

data DeleteAssessmentTargetResponse

See: deleteAssessmentTargetResponse smart constructor.