Network.AWS.Inspector.DescribeAssessmentTemplates

Description

Describes the assessment templates that are specified by the ARNs of the assessment templates.

Creating a Request

describeAssessmentTemplates

Arguments

:: NonEmpty Text
datAssessmentTemplateARNs

-> DescribeAssessmentTemplates

Creates a value of DescribeAssessmentTemplates with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

datAssessmentTemplateARNs - Undocumented member.

data DescribeAssessmentTemplates

See: describeAssessmentTemplates smart constructor.

Request Lenses

datAssessmentTemplateARNs :: Lens' DescribeAssessmentTemplates (NonEmpty Text)

Undocumented member.

Destructuring the Response

describeAssessmentTemplatesResponse

Arguments

:: Int
datrsResponseStatus

-> DescribeAssessmentTemplatesResponse

Creates a value of DescribeAssessmentTemplatesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

datrsResponseStatus - -- | The response status code.

datrsAssessmentTemplates - Information about the assessment templates.

datrsFailedItems - Assessment template details that cannot be described. An error code is provided for each failed item.

data DescribeAssessmentTemplatesResponse

See: describeAssessmentTemplatesResponse smart constructor.

Response Lenses

datrsResponseStatus :: Lens' DescribeAssessmentTemplatesResponse Int

- | The response status code.

datrsAssessmentTemplates :: Lens' DescribeAssessmentTemplatesResponse [AssessmentTemplate]

Information about the assessment templates.

datrsFailedItems :: Lens' DescribeAssessmentTemplatesResponse (HashMap Text FailedItemDetails)

Assessment template details that cannot be described. An error code is provided for each failed item.