Network.AWS.Inspector.DescribeCrossAccountAccessRole

Description

Describes the IAM role that enables Amazon Inspector to access your AWS account.

Creating a Request

describeCrossAccountAccessRole :: DescribeCrossAccountAccessRole

Creates a value of DescribeCrossAccountAccessRole with the minimum fields required to make a request.

data DescribeCrossAccountAccessRole

See: describeCrossAccountAccessRole smart constructor.

Destructuring the Response

describeCrossAccountAccessRoleResponse

Arguments

:: Int
dcaarrsResponseStatus

-> Text
dcaarrsRoleARN

-> Bool
dcaarrsValid

-> UTCTime
dcaarrsRegisteredAt

-> DescribeCrossAccountAccessRoleResponse

Creates a value of DescribeCrossAccountAccessRoleResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dcaarrsResponseStatus - -- | The response status code.

dcaarrsRoleARN - The ARN that specifies the IAM role that Amazon Inspector uses to access your AWS account.

dcaarrsValid - A Boolean value that specifies whether the IAM role has the necessary policies attached to enable Amazon Inspector to access your AWS account.

dcaarrsRegisteredAt - The date when the cross-account access role was registered.

data DescribeCrossAccountAccessRoleResponse

See: describeCrossAccountAccessRoleResponse smart constructor.

Response Lenses

dcaarrsResponseStatus :: Lens' DescribeCrossAccountAccessRoleResponse Int

- | The response status code.

dcaarrsRoleARN :: Lens' DescribeCrossAccountAccessRoleResponse Text

The ARN that specifies the IAM role that Amazon Inspector uses to access your AWS account.

dcaarrsValid :: Lens' DescribeCrossAccountAccessRoleResponse Bool

A Boolean value that specifies whether the IAM role has the necessary policies attached to enable Amazon Inspector to access your AWS account.

dcaarrsRegisteredAt :: Lens' DescribeCrossAccountAccessRoleResponse UTCTime

The date when the cross-account access role was registered.