Network.AWS.Inspector.DescribeFindings

Description

Describes the findings that are specified by the ARNs of the findings.

Creating a Request

describeFindings

Arguments

:: NonEmpty Text
dfFindingARNs

-> DescribeFindings

Creates a value of DescribeFindings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfLocale - The locale into which you want to translate a finding description, recommendation, and the short description that identifies the finding.

dfFindingARNs - The ARN that specifies the finding that you want to describe.

data DescribeFindings

See: describeFindings smart constructor.

Request Lenses

dfLocale :: Lens' DescribeFindings (Maybe Locale)

The locale into which you want to translate a finding description, recommendation, and the short description that identifies the finding.

dfFindingARNs :: Lens' DescribeFindings (NonEmpty Text)

The ARN that specifies the finding that you want to describe.

Destructuring the Response

describeFindingsResponse

Arguments

:: Int
dfrsResponseStatus

-> DescribeFindingsResponse

Creates a value of DescribeFindingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dfrsResponseStatus - -- | The response status code.

dfrsFindings - Information about the finding.

dfrsFailedItems - Finding details that cannot be described. An error code is provided for each failed item.

data DescribeFindingsResponse

See: describeFindingsResponse smart constructor.

Response Lenses

dfrsResponseStatus :: Lens' DescribeFindingsResponse Int

- | The response status code.

dfrsFindings :: Lens' DescribeFindingsResponse [Finding]

Information about the finding.

dfrsFailedItems :: Lens' DescribeFindingsResponse (HashMap Text FailedItemDetails)

Finding details that cannot be described. An error code is provided for each failed item.