Network.AWS.Inspector.DescribeResourceGroups

Description

Describes the resource groups that are specified by the ARNs of the resource groups.

Creating a Request

describeResourceGroups

Arguments

:: NonEmpty Text
drgResourceGroupARNs

-> DescribeResourceGroups

Creates a value of DescribeResourceGroups with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drgResourceGroupARNs - The ARN that specifies the resource group that you want to describe.

data DescribeResourceGroups

See: describeResourceGroups smart constructor.

Request Lenses

drgResourceGroupARNs :: Lens' DescribeResourceGroups (NonEmpty Text)

The ARN that specifies the resource group that you want to describe.

Destructuring the Response

describeResourceGroupsResponse

Arguments

:: Int
drgrsResponseStatus

-> DescribeResourceGroupsResponse

Creates a value of DescribeResourceGroupsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drgrsResponseStatus - -- | The response status code.

drgrsResourceGroups - Information about a resource group.

drgrsFailedItems - Resource group details that cannot be described. An error code is provided for each failed item.

data DescribeResourceGroupsResponse

See: describeResourceGroupsResponse smart constructor.

Response Lenses

drgrsResponseStatus :: Lens' DescribeResourceGroupsResponse Int

- | The response status code.

drgrsResourceGroups :: Lens' DescribeResourceGroupsResponse [ResourceGroup]

Information about a resource group.

drgrsFailedItems :: Lens' DescribeResourceGroupsResponse (HashMap Text FailedItemDetails)

Resource group details that cannot be described. An error code is provided for each failed item.