Network.AWS.Inspector.GetAssessmentReport

Description

Produces an assessment report that includes detailed and comprehensive results of a specified assessment run.

Creating a Request

getAssessmentReport

Arguments

:: Text
garAssessmentRunARN

-> ReportFileFormat
garReportFileFormat

-> ReportType
garReportType

-> GetAssessmentReport

Creates a value of GetAssessmentReport with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garAssessmentRunARN - The ARN that specifies the assessment run for which you want to generate a report.

garReportFileFormat - Specifies the file format (html or pdf) of the assessment report that you want to generate.

garReportType - Specifies the type of the assessment report that you want to generate. There are two types of assessment reports: a finding report and a full report. For more information, see Assessment Reports .

data GetAssessmentReport

See: getAssessmentReport smart constructor.

Request Lenses

garAssessmentRunARN :: Lens' GetAssessmentReport Text

The ARN that specifies the assessment run for which you want to generate a report.

garReportFileFormat :: Lens' GetAssessmentReport ReportFileFormat

Specifies the file format (html or pdf) of the assessment report that you want to generate.

garReportType :: Lens' GetAssessmentReport ReportType

Specifies the type of the assessment report that you want to generate. There are two types of assessment reports: a finding report and a full report. For more information, see Assessment Reports .

Destructuring the Response

getAssessmentReportResponse

Arguments

:: Int
garrsResponseStatus

-> ReportStatus
garrsStatus

-> GetAssessmentReportResponse

Creates a value of GetAssessmentReportResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

garrsUrl - Specifies the URL where you can find the generated assessment report. This parameter is only returned if the report is successfully generated.

garrsResponseStatus - -- | The response status code.

garrsStatus - Specifies the status of the request to generate an assessment report.

data GetAssessmentReportResponse

See: getAssessmentReportResponse smart constructor.

Response Lenses

garrsUrl :: Lens' GetAssessmentReportResponse (Maybe Text)

Specifies the URL where you can find the generated assessment report. This parameter is only returned if the report is successfully generated.

garrsResponseStatus :: Lens' GetAssessmentReportResponse Int

- | The response status code.

garrsStatus :: Lens' GetAssessmentReportResponse ReportStatus

Specifies the status of the request to generate an assessment report.