Network.AWS.Inspector.ListAssessmentRuns

Description

Lists the assessment runs that correspond to the assessment templates that are specified by the ARNs of the assessment templates.

Creating a Request

listAssessmentRuns :: ListAssessmentRuns

Creates a value of ListAssessmentRuns with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larNextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRuns action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

larFilter - You can use this parameter to specify a subset of data to be included in the action's response. For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

larAssessmentTemplateARNs - The ARNs that specify the assessment templates whose assessment runs you want to list.

larMaxResults - You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

data ListAssessmentRuns

See: listAssessmentRuns smart constructor.

Request Lenses

larNextToken :: Lens' ListAssessmentRuns (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentRuns action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

larFilter :: Lens' ListAssessmentRuns (Maybe AssessmentRunFilter)

You can use this parameter to specify a subset of data to be included in the action's response. For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

larAssessmentTemplateARNs :: Lens' ListAssessmentRuns [Text]

The ARNs that specify the assessment templates whose assessment runs you want to list.

larMaxResults :: Lens' ListAssessmentRuns (Maybe Int)

You can use this parameter to indicate the maximum number of items that you want in the response. The default value is 10. The maximum value is 500.

Destructuring the Response

listAssessmentRunsResponse

Arguments

:: Int
larrsResponseStatus

-> ListAssessmentRunsResponse

Creates a value of ListAssessmentRunsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

larrsNextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

larrsResponseStatus - -- | The response status code.

larrsAssessmentRunARNs - A list of ARNs that specifies the assessment runs that are returned by the action.

data ListAssessmentRunsResponse

See: listAssessmentRunsResponse smart constructor.

Response Lenses

larrsNextToken :: Lens' ListAssessmentRunsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

larrsResponseStatus :: Lens' ListAssessmentRunsResponse Int

- | The response status code.

larrsAssessmentRunARNs :: Lens' ListAssessmentRunsResponse [Text]

A list of ARNs that specifies the assessment runs that are returned by the action.