Network.AWS.Inspector.ListAssessmentTargets

Description

Lists the ARNs of the assessment targets within this AWS account. For more information about assessment targets, see Amazon Inspector Assessment Targets .

Creating a Request

listAssessmentTargets :: ListAssessmentTargets

Creates a value of ListAssessmentTargets with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lNextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTargets action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

lFilter - You can use this parameter to specify a subset of data to be included in the action's response. For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

lMaxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

data ListAssessmentTargets

See: listAssessmentTargets smart constructor.

Request Lenses

lNextToken :: Lens' ListAssessmentTargets (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListAssessmentTargets action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

lFilter :: Lens' ListAssessmentTargets (Maybe AssessmentTargetFilter)

You can use this parameter to specify a subset of data to be included in the action's response. For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

lMaxResults :: Lens' ListAssessmentTargets (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Destructuring the Response

listAssessmentTargetsResponse

Arguments

:: Int
lrsResponseStatus

-> ListAssessmentTargetsResponse

Creates a value of ListAssessmentTargetsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrsNextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

lrsResponseStatus - -- | The response status code.

lrsAssessmentTargetARNs - A list of ARNs that specifies the assessment targets that are returned by the action.

data ListAssessmentTargetsResponse

See: listAssessmentTargetsResponse smart constructor.

Response Lenses

lrsNextToken :: Lens' ListAssessmentTargetsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

lrsResponseStatus :: Lens' ListAssessmentTargetsResponse Int

- | The response status code.

lrsAssessmentTargetARNs :: Lens' ListAssessmentTargetsResponse [Text]

A list of ARNs that specifies the assessment targets that are returned by the action.