Network.AWS.Inspector.ListFindings

Description

Lists findings that are generated by the assessment runs that are specified by the ARNs of the assessment runs.

Creating a Request

listFindings :: ListFindings

Creates a value of ListFindings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfAssessmentRunARNs - The ARNs of the assessment runs that generate the findings that you want to list.

lfNextToken - You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListFindings action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

lfFilter - You can use this parameter to specify a subset of data to be included in the action's response. For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

lfMaxResults - You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

data ListFindings

See: listFindings smart constructor.

Request Lenses

lfAssessmentRunARNs :: Lens' ListFindings [Text]

The ARNs of the assessment runs that generate the findings that you want to list.

lfNextToken :: Lens' ListFindings (Maybe Text)

You can use this parameter when paginating results. Set the value of this parameter to null on your first call to the ListFindings action. Subsequent calls to the action fill nextToken in the request with the value of NextToken from the previous response to continue listing data.

lfFilter :: Lens' ListFindings (Maybe FindingFilter)

You can use this parameter to specify a subset of data to be included in the action's response. For a record to match a filter, all specified filter attributes must match. When multiple values are specified for a filter attribute, any of the values can match.

lfMaxResults :: Lens' ListFindings (Maybe Int)

You can use this parameter to indicate the maximum number of items you want in the response. The default value is 10. The maximum value is 500.

Destructuring the Response

listFindingsResponse

Arguments

:: Int
lfrsResponseStatus

-> ListFindingsResponse

Creates a value of ListFindingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lfrsNextToken - When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

lfrsResponseStatus - -- | The response status code.

lfrsFindingARNs - A list of ARNs that specifies the findings returned by the action.

data ListFindingsResponse

See: listFindingsResponse smart constructor.

Response Lenses

lfrsNextToken :: Lens' ListFindingsResponse (Maybe Text)

When a response is generated, if there is more data to be listed, this parameter is present in the response and contains the value to use for the nextToken parameter in a subsequent pagination request. If there is no more data to be listed, this parameter is set to null.

lfrsResponseStatus :: Lens' ListFindingsResponse Int

- | The response status code.

lfrsFindingARNs :: Lens' ListFindingsResponse [Text]

A list of ARNs that specifies the findings returned by the action.