Network.AWS.Inspector.RemoveAttributesFromFindings

Description

Removes entire attributes (key and value pairs) from the findings that are specified by the ARNs of the findings where an attribute with the specified key exists.

Creating a Request

removeAttributesFromFindings

Arguments

:: NonEmpty Text
raffFindingARNs

-> RemoveAttributesFromFindings

Creates a value of RemoveAttributesFromFindings with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

raffFindingARNs - The ARNs that specify the findings that you want to remove attributes from.

raffAttributeKeys - The array of attribute keys that you want to remove from specified findings.

data RemoveAttributesFromFindings

See: removeAttributesFromFindings smart constructor.

Request Lenses

raffFindingARNs :: Lens' RemoveAttributesFromFindings (NonEmpty Text)

The ARNs that specify the findings that you want to remove attributes from.

raffAttributeKeys :: Lens' RemoveAttributesFromFindings [Text]

The array of attribute keys that you want to remove from specified findings.

Destructuring the Response

removeAttributesFromFindingsResponse

Arguments

:: Int
raffrsResponseStatus

-> RemoveAttributesFromFindingsResponse

Creates a value of RemoveAttributesFromFindingsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

raffrsResponseStatus - -- | The response status code.

raffrsFailedItems - Attributes details that cannot be described. An error code is provided for each failed item.

data RemoveAttributesFromFindingsResponse

See: removeAttributesFromFindingsResponse smart constructor.

Response Lenses

raffrsResponseStatus :: Lens' RemoveAttributesFromFindingsResponse Int

- | The response status code.

raffrsFailedItems :: Lens' RemoveAttributesFromFindingsResponse (HashMap Text FailedItemDetails)

Attributes details that cannot be described. An error code is provided for each failed item.