Network.AWS.Inspector.SubscribeToEvent

Description

Enables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

Creating a Request

subscribeToEvent

Arguments

:: Text
steResourceARN

-> InspectorEvent
steEvent

-> Text
steTopicARN

-> SubscribeToEvent

Creates a value of SubscribeToEvent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

steResourceARN - The ARN of the assessment template that is used during the event for which you want to receive SNS notifications.

steEvent - The event for which you want to receive SNS notifications.

steTopicARN - The ARN of the SNS topic to which the SNS notifications are sent.

data SubscribeToEvent

See: subscribeToEvent smart constructor.

Request Lenses

steResourceARN :: Lens' SubscribeToEvent Text

The ARN of the assessment template that is used during the event for which you want to receive SNS notifications.

steEvent :: Lens' SubscribeToEvent InspectorEvent

The event for which you want to receive SNS notifications.

steTopicARN :: Lens' SubscribeToEvent Text

The ARN of the SNS topic to which the SNS notifications are sent.

Destructuring the Response

subscribeToEventResponse :: SubscribeToEventResponse

Creates a value of SubscribeToEventResponse with the minimum fields required to make a request.

data SubscribeToEventResponse

See: subscribeToEventResponse smart constructor.