Network.AWS.Inspector.Types

Description

Service Configuration

inspector :: Service

API version 2016-02-16 of the Amazon Inspector SDK configuration.

Errors

_AccessDeniedException :: AsError a => Getting (First ServiceError) a ServiceError

You do not have required permissions to access the requested resource.

_AssessmentRunInProgressException :: AsError a => Getting (First ServiceError) a ServiceError

You cannot perform a specified action if an assessment run is currently in progress.

_NoSuchEntityException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it referenced an entity that does not exist. The error code describes the entity.

_UnsupportedFeatureException :: AsError a => Getting (First ServiceError) a ServiceError

Used by the GetAssessmentReport API. The request was rejected because you tried to generate a report for an assessment run that existed before reporting was supported in Amazon Inspector. You can only generate reports for assessment runs that took place or will take place after generating reports in Amazon Inspector became available.

_AgentsAlreadyRunningAssessmentException :: AsError a => Getting (First ServiceError) a ServiceError

You started an assessment run, but one of the instances is already participating in another assessment run.

_InvalidCrossAccountRoleException :: AsError a => Getting (First ServiceError) a ServiceError

Amazon Inspector cannot assume the cross-account role that it needs to list your EC2 instances during the assessment run.

_InvalidInputException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because an invalid or out-of-range value was supplied for an input parameter.

_InternalException :: AsError a => Getting (First ServiceError) a ServiceError

Internal server error.

_LimitExceededException :: AsError a => Getting (First ServiceError) a ServiceError

The request was rejected because it attempted to create resources beyond the current AWS account limits. The error code describes the limit exceeded.

AgentHealth

data AgentHealth

Constructors

AHHealthy

AHUnhealthy

AgentHealthCode

data AgentHealthCode

Constructors

Idle

Running

Shutdown

Throttled

Unhealthy

Unknown

AssessmentRunNotificationSNSStatusCode

data AssessmentRunNotificationSNSStatusCode

Constructors

AccessDenied

InternalError

Success

TopicDoesNotExist

AssessmentRunState

data AssessmentRunState

Constructors

Canceled

CollectingData

Completed

CompletedWithErrors

Created

DataCollected

Error'

EvaluatingRules

Failed

StartDataCollectionInProgress

StartDataCollectionPending

StartEvaluatingRulesPending

StopDataCollectionPending

AssetType

data AssetType

Constructors

EC2Instance

FailedItemErrorCode

data FailedItemErrorCode

Constructors

FIECAccessDenied

FIECDuplicateARN

FIECInternalError

FIECInvalidARN

FIECItemDoesNotExist

FIECLimitExceeded

InspectorEvent

data InspectorEvent

Constructors

AssessmentRunCompleted

AssessmentRunStarted

AssessmentRunStateChanged

FindingReported

Other

Locale

data Locale

Constructors

EnUs

ReportFileFormat

data ReportFileFormat

Constructors

HTML

Pdf

ReportStatus

data ReportStatus

Constructors

RSCompleted

RSFailed

RSWorkInProgress

ReportType

data ReportType

Constructors

Finding

Full

Severity

data Severity

Constructors

High

Informational

Low

Medium

Undefined

StopAction

data StopAction

Constructors

SkipEvaluation

StartEvaluation

AgentFilter

data AgentFilter

Contains information about an Amazon Inspector agent. This data type is used as a request parameter in the ListAssessmentRunAgents action.

See: agentFilter smart constructor.

agentFilter :: AgentFilter

Creates a value of AgentFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

afAgentHealths - The current health state of the agent. Values can be set to HEALTHY or UNHEALTHY .

afAgentHealthCodes - The detailed health state of the agent. Values can be set to IDLE , RUNNING , SHUTDOWN , UNHEALTHY , THROTTLED , and UNKNOWN .

afAgentHealths :: Lens' AgentFilter [AgentHealth]

The current health state of the agent. Values can be set to HEALTHY or UNHEALTHY .

afAgentHealthCodes :: Lens' AgentFilter [AgentHealthCode]

The detailed health state of the agent. Values can be set to IDLE , RUNNING , SHUTDOWN , UNHEALTHY , THROTTLED , and UNKNOWN .

AgentPreview

data AgentPreview

Used as a response element in the PreviewAgents action.

See: agentPreview smart constructor.

agentPreview

Arguments

:: Text
apAgentId

-> AgentPreview

Creates a value of AgentPreview with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

apAutoScalingGroup - The Auto Scaling group for the EC2 instance where the agent is installed.

apAgentId - The ID of the EC2 instance where the agent is installed.

apAutoScalingGroup :: Lens' AgentPreview (Maybe Text)

The Auto Scaling group for the EC2 instance where the agent is installed.

apAgentId :: Lens' AgentPreview Text

The ID of the EC2 instance where the agent is installed.

AssessmentRun

data AssessmentRun

A snapshot of an Amazon Inspector assessment run that contains the findings of the assessment run .

Used as the response element in the DescribeAssessmentRuns action.

See: assessmentRun smart constructor.

assessmentRun

Arguments

:: Text
arArn

-> Text
arName

-> Text
arAssessmentTemplateARN

-> AssessmentRunState
arState

-> Natural
arDurationInSeconds

-> NonEmpty Text
arRulesPackageARNs

-> UTCTime
arCreatedAt

-> UTCTime
arStateChangedAt

-> Bool
arDataCollected

-> AssessmentRun

Creates a value of AssessmentRun with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arStartedAt - The time when StartAssessmentRun was called.

arCompletedAt - The assessment run completion time that corresponds to the rules packages evaluation completion time or failure.

arArn - The ARN of the assessment run.

arName - The auto-generated name for the assessment run.

arAssessmentTemplateARN - The ARN of the assessment template that is associated with the assessment run.

arState - The state of the assessment run.

arDurationInSeconds - The duration of the assessment run.

arRulesPackageARNs - The rules packages selected for the assessment run.

arUserAttributesForFindings - The user-defined attributes that are assigned to every generated finding.

arCreatedAt - The time when StartAssessmentRun was called.

arStateChangedAt - The last time when the assessment run's state changed.

arDataCollected - A Boolean value (true or false) that specifies whether the process of collecting data from the agents is completed.

arStateChanges - A list of the assessment run state changes.

arNotifications - A list of notifications for the event subscriptions. A notification about a particular generated finding is added to this list only once.

arFindingCounts - Provides a total count of generated findings per severity.

arStartedAt :: Lens' AssessmentRun (Maybe UTCTime)

The time when StartAssessmentRun was called.

arCompletedAt :: Lens' AssessmentRun (Maybe UTCTime)

The assessment run completion time that corresponds to the rules packages evaluation completion time or failure.

arArn :: Lens' AssessmentRun Text

The ARN of the assessment run.

arName :: Lens' AssessmentRun Text

The auto-generated name for the assessment run.

arAssessmentTemplateARN :: Lens' AssessmentRun Text

The ARN of the assessment template that is associated with the assessment run.

arState :: Lens' AssessmentRun AssessmentRunState

The state of the assessment run.

arDurationInSeconds :: Lens' AssessmentRun Natural

The duration of the assessment run.

arRulesPackageARNs :: Lens' AssessmentRun (NonEmpty Text)

The rules packages selected for the assessment run.

arUserAttributesForFindings :: Lens' AssessmentRun [Attribute]

The user-defined attributes that are assigned to every generated finding.

arCreatedAt :: Lens' AssessmentRun UTCTime

The time when StartAssessmentRun was called.

arStateChangedAt :: Lens' AssessmentRun UTCTime

The last time when the assessment run's state changed.

arDataCollected :: Lens' AssessmentRun Bool

A Boolean value (true or false) that specifies whether the process of collecting data from the agents is completed.

arStateChanges :: Lens' AssessmentRun [AssessmentRunStateChange]

A list of the assessment run state changes.

arNotifications :: Lens' AssessmentRun [AssessmentRunNotification]

A list of notifications for the event subscriptions. A notification about a particular generated finding is added to this list only once.

arFindingCounts :: Lens' AssessmentRun (HashMap Severity Int)

Provides a total count of generated findings per severity.

AssessmentRunAgent

data AssessmentRunAgent

Contains information about an Amazon Inspector agent. This data type is used as a response element in the ListAssessmentRunAgents action.

See: assessmentRunAgent smart constructor.

assessmentRunAgent

Arguments

:: Text
araAgentId

-> Text
araAssessmentRunARN

-> AgentHealth
araAgentHealth

-> AgentHealthCode
araAgentHealthCode

-> AssessmentRunAgent

Creates a value of AssessmentRunAgent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

araAutoScalingGroup - The Auto Scaling group of the EC2 instance that is specified by the agent ID.

araAgentHealthDetails - The description for the agent health code.

araAgentId - The AWS account of the EC2 instance where the agent is installed.

araAssessmentRunARN - The ARN of the assessment run that is associated with the agent.

araAgentHealth - The current health state of the agent.

araAgentHealthCode - The detailed health state of the agent.

araTelemetryMetadata - The Amazon Inspector application data metrics that are collected by the agent.

araAutoScalingGroup :: Lens' AssessmentRunAgent (Maybe Text)

The Auto Scaling group of the EC2 instance that is specified by the agent ID.

araAgentHealthDetails :: Lens' AssessmentRunAgent (Maybe Text)

The description for the agent health code.

araAgentId :: Lens' AssessmentRunAgent Text

The AWS account of the EC2 instance where the agent is installed.

araAssessmentRunARN :: Lens' AssessmentRunAgent Text

The ARN of the assessment run that is associated with the agent.

araAgentHealth :: Lens' AssessmentRunAgent AgentHealth

The current health state of the agent.

araAgentHealthCode :: Lens' AssessmentRunAgent AgentHealthCode

The detailed health state of the agent.

araTelemetryMetadata :: Lens' AssessmentRunAgent [TelemetryMetadata]

The Amazon Inspector application data metrics that are collected by the agent.

AssessmentRunFilter

data AssessmentRunFilter

Used as the request parameter in the ListAssessmentRuns action.

See: assessmentRunFilter smart constructor.

assessmentRunFilter :: AssessmentRunFilter

Creates a value of AssessmentRunFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arfStates - For a record to match a filter, one of the values specified for this data type property must be the exact match of the value of the assessmentRunState property of the AssessmentRun data type.

arfNamePattern - For a record to match a filter, an explicit value or a string containing a wildcard that is specified for this data type property must match the value of the assessmentRunName property of the AssessmentRun data type.

arfStartTimeRange - For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the startTime property of the AssessmentRun data type.

arfStateChangeTimeRange - For a record to match a filter, the value that is specified for this data type property must match the stateChangedAt property of the AssessmentRun data type.

arfRulesPackageARNs - For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the rulesPackages property of the AssessmentRun data type.

arfCompletionTimeRange - For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the completedAt property of the AssessmentRun data type.

arfDurationRange - For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentRun data type.

arfStates :: Lens' AssessmentRunFilter [AssessmentRunState]

For a record to match a filter, one of the values specified for this data type property must be the exact match of the value of the assessmentRunState property of the AssessmentRun data type.

arfNamePattern :: Lens' AssessmentRunFilter (Maybe Text)

For a record to match a filter, an explicit value or a string containing a wildcard that is specified for this data type property must match the value of the assessmentRunName property of the AssessmentRun data type.

arfStartTimeRange :: Lens' AssessmentRunFilter (Maybe TimestampRange)

For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the startTime property of the AssessmentRun data type.

arfStateChangeTimeRange :: Lens' AssessmentRunFilter (Maybe TimestampRange)

For a record to match a filter, the value that is specified for this data type property must match the stateChangedAt property of the AssessmentRun data type.

arfRulesPackageARNs :: Lens' AssessmentRunFilter [Text]

For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the rulesPackages property of the AssessmentRun data type.

arfCompletionTimeRange :: Lens' AssessmentRunFilter (Maybe TimestampRange)

For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the completedAt property of the AssessmentRun data type.

arfDurationRange :: Lens' AssessmentRunFilter (Maybe DurationRange)

For a record to match a filter, the value that is specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentRun data type.

AssessmentRunNotification

data AssessmentRunNotification

Used as one of the elements of the AssessmentRun data type.

See: assessmentRunNotification smart constructor.

assessmentRunNotification

Arguments

:: UTCTime
arnDate

-> InspectorEvent
arnEvent

-> Bool
arnError

-> AssessmentRunNotification

Creates a value of AssessmentRunNotification with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arnSnsTopicARN - The SNS topic to which the SNS notification is sent.

arnSnsPublishStatusCode - The status code of the SNS notification.

arnMessage - The message included in the notification.

arnDate - The date of the notification.

arnEvent - The event for which a notification is sent.

arnError - The Boolean value that specifies whether the notification represents an error.

arnSnsTopicARN :: Lens' AssessmentRunNotification (Maybe Text)

The SNS topic to which the SNS notification is sent.

arnSnsPublishStatusCode :: Lens' AssessmentRunNotification (Maybe AssessmentRunNotificationSNSStatusCode)

The status code of the SNS notification.

arnMessage :: Lens' AssessmentRunNotification (Maybe Text)

The message included in the notification.

arnDate :: Lens' AssessmentRunNotification UTCTime

The date of the notification.

arnEvent :: Lens' AssessmentRunNotification InspectorEvent

The event for which a notification is sent.

arnError :: Lens' AssessmentRunNotification Bool

The Boolean value that specifies whether the notification represents an error.

AssessmentRunStateChange

data AssessmentRunStateChange

Used as one of the elements of the AssessmentRun data type.

See: assessmentRunStateChange smart constructor.

assessmentRunStateChange

Arguments

:: UTCTime
arscStateChangedAt

-> AssessmentRunState
arscState

-> AssessmentRunStateChange

Creates a value of AssessmentRunStateChange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

arscStateChangedAt - The last time the assessment run state changed.

arscState - The assessment run state.

arscStateChangedAt :: Lens' AssessmentRunStateChange UTCTime

The last time the assessment run state changed.

arscState :: Lens' AssessmentRunStateChange AssessmentRunState

The assessment run state.

AssessmentTarget

data AssessmentTarget

Contains information about an Amazon Inspector application. This data type is used as the response element in the DescribeAssessmentTargets action.

See: assessmentTarget smart constructor.

assessmentTarget

Arguments

:: Text
aArn

-> Text
aName

-> Text
aResourceGroupARN

-> UTCTime
aCreatedAt

-> UTCTime
aUpdatedAt

-> AssessmentTarget

Creates a value of AssessmentTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aArn - The ARN that specifies the Amazon Inspector assessment target.

aName - The name of the Amazon Inspector assessment target.

aResourceGroupARN - The ARN that specifies the resource group that is associated with the assessment target.

aCreatedAt - The time at which the assessment target is created.

aUpdatedAt - The time at which UpdateAssessmentTarget is called.

aArn :: Lens' AssessmentTarget Text

The ARN that specifies the Amazon Inspector assessment target.

aName :: Lens' AssessmentTarget Text

The name of the Amazon Inspector assessment target.

aResourceGroupARN :: Lens' AssessmentTarget Text

The ARN that specifies the resource group that is associated with the assessment target.

aCreatedAt :: Lens' AssessmentTarget UTCTime

The time at which the assessment target is created.

aUpdatedAt :: Lens' AssessmentTarget UTCTime

The time at which UpdateAssessmentTarget is called.

AssessmentTargetFilter

data AssessmentTargetFilter

Used as the request parameter in the ListAssessmentTargets action.

See: assessmentTargetFilter smart constructor.

assessmentTargetFilter :: AssessmentTargetFilter

Creates a value of AssessmentTargetFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atfAssessmentTargetNamePattern - For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTargetName property of the AssessmentTarget data type.

atfAssessmentTargetNamePattern :: Lens' AssessmentTargetFilter (Maybe Text)

For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTargetName property of the AssessmentTarget data type.

AssessmentTemplate

data AssessmentTemplate

Contains information about an Amazon Inspector assessment template. This data type is used as the response element in the DescribeAssessmentTemplates action.

See: assessmentTemplate smart constructor.

assessmentTemplate

Arguments

:: Text
atArn

-> Text
atName

-> Text
atAssessmentTargetARN

-> Natural
atDurationInSeconds

-> UTCTime
atCreatedAt

-> AssessmentTemplate

Creates a value of AssessmentTemplate with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atArn - The ARN of the assessment template.

atName - The name of the assessment template.

atAssessmentTargetARN - The ARN of the assessment target that corresponds to this assessment template.

atDurationInSeconds - The duration in seconds specified for this assessment tempate. The default value is 3600 seconds (one hour). The maximum value is 86400 seconds (one day).

atRulesPackageARNs - The rules packages that are specified for this assessment template.

atUserAttributesForFindings - The user-defined attributes that are assigned to every generated finding from the assessment run that uses this assessment template.

atCreatedAt - The time at which the assessment template is created.

atArn :: Lens' AssessmentTemplate Text

The ARN of the assessment template.

atName :: Lens' AssessmentTemplate Text

The name of the assessment template.

atAssessmentTargetARN :: Lens' AssessmentTemplate Text

The ARN of the assessment target that corresponds to this assessment template.

atDurationInSeconds :: Lens' AssessmentTemplate Natural

The duration in seconds specified for this assessment tempate. The default value is 3600 seconds (one hour). The maximum value is 86400 seconds (one day).

atRulesPackageARNs :: Lens' AssessmentTemplate [Text]

The rules packages that are specified for this assessment template.

atUserAttributesForFindings :: Lens' AssessmentTemplate [Attribute]

The user-defined attributes that are assigned to every generated finding from the assessment run that uses this assessment template.

atCreatedAt :: Lens' AssessmentTemplate UTCTime

The time at which the assessment template is created.

AssessmentTemplateFilter

data AssessmentTemplateFilter

Used as the request parameter in the ListAssessmentTemplates action.

See: assessmentTemplateFilter smart constructor.

assessmentTemplateFilter :: AssessmentTemplateFilter

Creates a value of AssessmentTemplateFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

atfNamePattern - For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTemplateName property of the AssessmentTemplate data type.

atfRulesPackageARNs - For a record to match a filter, the values that are specified for this data type property must be contained in the list of values of the rulesPackageArns property of the AssessmentTemplate data type.

atfDurationRange - For a record to match a filter, the value specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentTemplate data type.

atfNamePattern :: Lens' AssessmentTemplateFilter (Maybe Text)

For a record to match a filter, an explicit value or a string that contains a wildcard that is specified for this data type property must match the value of the assessmentTemplateName property of the AssessmentTemplate data type.

atfRulesPackageARNs :: Lens' AssessmentTemplateFilter [Text]

For a record to match a filter, the values that are specified for this data type property must be contained in the list of values of the rulesPackageArns property of the AssessmentTemplate data type.

atfDurationRange :: Lens' AssessmentTemplateFilter (Maybe DurationRange)

For a record to match a filter, the value specified for this data type property must inclusively match any value between the specified minimum and maximum values of the durationInSeconds property of the AssessmentTemplate data type.

AssetAttributes

data AssetAttributes

A collection of attributes of the host from which the finding is generated.

See: assetAttributes smart constructor.

assetAttributes

Arguments

:: Natural
aaSchemaVersion

-> AssetAttributes

Creates a value of AssetAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aaHostname - The hostname of the EC2 instance where the finding is generated.

aaAutoScalingGroup - The Auto Scaling group of the EC2 instance where the finding is generated.

aaIpv4Addresses - The list of IP v4 addresses of the EC2 instance where the finding is generated.

aaAgentId - The ID of the agent that is installed on the EC2 instance where the finding is generated.

aaAmiId - The ID of the Amazon Machine Image (AMI) that is installed on the EC2 instance where the finding is generated.

aaSchemaVersion - The schema version of this data type.

aaHostname :: Lens' AssetAttributes (Maybe Text)

The hostname of the EC2 instance where the finding is generated.

aaAutoScalingGroup :: Lens' AssetAttributes (Maybe Text)

The Auto Scaling group of the EC2 instance where the finding is generated.

aaIpv4Addresses :: Lens' AssetAttributes [Text]

The list of IP v4 addresses of the EC2 instance where the finding is generated.

aaAgentId :: Lens' AssetAttributes (Maybe Text)

The ID of the agent that is installed on the EC2 instance where the finding is generated.

aaAmiId :: Lens' AssetAttributes (Maybe Text)

The ID of the Amazon Machine Image (AMI) that is installed on the EC2 instance where the finding is generated.

aaSchemaVersion :: Lens' AssetAttributes Natural

The schema version of this data type.

Attribute

data Attribute

This data type is used as a request parameter in the AddAttributesToFindings and CreateAssessmentTemplate actions.

See: attribute smart constructor.

attribute

Arguments

:: Text
aKey

-> Attribute

Creates a value of Attribute with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

aValue - The value assigned to the attribute key.

aKey - The attribute key.

aValue :: Lens' Attribute (Maybe Text)

The value assigned to the attribute key.

aKey :: Lens' Attribute Text

The attribute key.

DurationRange

data DurationRange

This data type is used in the AssessmentTemplateFilter data type.

See: durationRange smart constructor.

durationRange :: DurationRange

Creates a value of DurationRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drMinSeconds - The minimum value of the duration range. Must be greater than zero.

drMaxSeconds - The maximum value of the duration range. Must be less than or equal to 604800 seconds (1 week).

drMinSeconds :: Lens' DurationRange (Maybe Natural)

The minimum value of the duration range. Must be greater than zero.

drMaxSeconds :: Lens' DurationRange (Maybe Natural)

The maximum value of the duration range. Must be less than or equal to 604800 seconds (1 week).

EventSubscription

data EventSubscription

This data type is used in the Subscription data type.

See: eventSubscription smart constructor.

eventSubscription

Arguments

:: InspectorEvent
esEvent

-> UTCTime
esSubscribedAt

-> EventSubscription

Creates a value of EventSubscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

esEvent - The event for which Amazon Simple Notification Service (SNS) notifications are sent.

esSubscribedAt - The time at which SubscribeToEvent is called.

esEvent :: Lens' EventSubscription InspectorEvent

The event for which Amazon Simple Notification Service (SNS) notifications are sent.

esSubscribedAt :: Lens' EventSubscription UTCTime

The time at which SubscribeToEvent is called.

FailedItemDetails

data FailedItemDetails

Includes details about the failed items.

See: failedItemDetails smart constructor.

failedItemDetails

Arguments

:: FailedItemErrorCode
fidFailureCode

-> Bool
fidRetryable

-> FailedItemDetails

Creates a value of FailedItemDetails with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fidFailureCode - The status code of a failed item.

fidRetryable - Indicates whether you can immediately retry a request for this item for a specified resource.

fidFailureCode :: Lens' FailedItemDetails FailedItemErrorCode

The status code of a failed item.

fidRetryable :: Lens' FailedItemDetails Bool

Indicates whether you can immediately retry a request for this item for a specified resource.

Finding

data Finding

Contains information about an Amazon Inspector finding. This data type is used as the response element in the DescribeFindings action.

See: finding smart constructor.

finding

Arguments

:: Text
fArn

-> UTCTime
fCreatedAt

-> UTCTime
fUpdatedAt

-> Finding

Creates a value of Finding with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

fService - The data element is set to Inspector.

fSeverity - The finding severity. Values can be set to High, Medium, Low, and Informational.

fSchemaVersion - The schema version of this data type.

fConfidence - This data element is currently not used.

fAssetAttributes - A collection of attributes of the host from which the finding is generated.

fServiceAttributes - This data type is used in the Finding data type.

fId - The ID of the finding.

fNumericSeverity - The numeric value of the finding severity.

fAssetType - The type of the host from which the finding is generated.

fTitle - The name of the finding.

fIndicatorOfCompromise - This data element is currently not used.

fDescription - The description of the finding.

fRecommendation - The recommendation for the finding.

fArn - The ARN that specifies the finding.

fAttributes - The system-defined attributes for the finding.

fUserAttributes - The user-defined attributes that are assigned to the finding.

fCreatedAt - The time when the finding was generated.

fUpdatedAt - The time when AddAttributesToFindings is called.

fService :: Lens' Finding (Maybe Text)

The data element is set to Inspector.

fSeverity :: Lens' Finding (Maybe Severity)

The finding severity. Values can be set to High, Medium, Low, and Informational.

fSchemaVersion :: Lens' Finding (Maybe Natural)

The schema version of this data type.

fConfidence :: Lens' Finding (Maybe Natural)

This data element is currently not used.

fAssetAttributes :: Lens' Finding (Maybe AssetAttributes)

A collection of attributes of the host from which the finding is generated.

fServiceAttributes :: Lens' Finding (Maybe InspectorServiceAttributes)

This data type is used in the Finding data type.

fId :: Lens' Finding (Maybe Text)

The ID of the finding.

fNumericSeverity :: Lens' Finding (Maybe Double)

The numeric value of the finding severity.

fAssetType :: Lens' Finding (Maybe AssetType)

The type of the host from which the finding is generated.

fTitle :: Lens' Finding (Maybe Text)

The name of the finding.

fIndicatorOfCompromise :: Lens' Finding (Maybe Bool)

This data element is currently not used.

fDescription :: Lens' Finding (Maybe Text)

The description of the finding.

fRecommendation :: Lens' Finding (Maybe Text)

The recommendation for the finding.

fArn :: Lens' Finding Text

The ARN that specifies the finding.

fAttributes :: Lens' Finding [Attribute]

The system-defined attributes for the finding.

fUserAttributes :: Lens' Finding [Attribute]

The user-defined attributes that are assigned to the finding.

fCreatedAt :: Lens' Finding UTCTime

The time when the finding was generated.

fUpdatedAt :: Lens' Finding UTCTime

The time when AddAttributesToFindings is called.

FindingFilter

data FindingFilter

This data type is used as a request parameter in the ListFindings action.

See: findingFilter smart constructor.

findingFilter :: FindingFilter

Creates a value of FindingFilter with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ffAgentIds - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the agentId property of the Finding data type.

ffRuleNames - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the ruleName property of the Finding data type.

ffUserAttributes - For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the userAttributes property of the Finding data type.

ffRulesPackageARNs - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the rulesPackageArn property of the Finding data type.

ffAttributes - For a record to match a filter, the list of values that are specified for this data type property must be contained in the list of values of the attributes property of the Finding data type.

ffSeverities - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the severity property of the Finding data type.

ffCreationTimeRange - The time range during which the finding is generated.

ffAutoScalingGroups - For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the autoScalingGroup property of the Finding data type.

ffAgentIds :: Lens' FindingFilter [Text]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the agentId property of the Finding data type.

ffRuleNames :: Lens' FindingFilter [Text]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the ruleName property of the Finding data type.

ffUserAttributes :: Lens' FindingFilter [Attribute]

For a record to match a filter, the value that is specified for this data type property must be contained in the list of values of the userAttributes property of the Finding data type.

ffRulesPackageARNs :: Lens' FindingFilter [Text]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the rulesPackageArn property of the Finding data type.

ffAttributes :: Lens' FindingFilter [Attribute]

For a record to match a filter, the list of values that are specified for this data type property must be contained in the list of values of the attributes property of the Finding data type.

ffSeverities :: Lens' FindingFilter [Severity]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the severity property of the Finding data type.

ffCreationTimeRange :: Lens' FindingFilter (Maybe TimestampRange)

The time range during which the finding is generated.

ffAutoScalingGroups :: Lens' FindingFilter [Text]

For a record to match a filter, one of the values that is specified for this data type property must be the exact match of the value of the autoScalingGroup property of the Finding data type.

InspectorServiceAttributes

data InspectorServiceAttributes

This data type is used in the Finding data type.

See: inspectorServiceAttributes smart constructor.

inspectorServiceAttributes

Arguments

:: Natural
isaSchemaVersion

-> InspectorServiceAttributes

Creates a value of InspectorServiceAttributes with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

isaRulesPackageARN - The ARN of the rules package that is used to generate the finding.

isaAssessmentRunARN - The ARN of the assessment run during which the finding is generated.

isaSchemaVersion - The schema version of this data type.

isaRulesPackageARN :: Lens' InspectorServiceAttributes (Maybe Text)

The ARN of the rules package that is used to generate the finding.

isaAssessmentRunARN :: Lens' InspectorServiceAttributes (Maybe Text)

The ARN of the assessment run during which the finding is generated.

isaSchemaVersion :: Lens' InspectorServiceAttributes Natural

The schema version of this data type.

ResourceGroup

data ResourceGroup

Contains information about a resource group. The resource group defines a set of tags that, when queried, identify the AWS resources that make up the assessment target. This data type is used as the response element in the DescribeResourceGroups action.

See: resourceGroup smart constructor.

resourceGroup

Arguments

:: Text
rgArn

-> NonEmpty ResourceGroupTag
rgTags

-> UTCTime
rgCreatedAt

-> ResourceGroup

Creates a value of ResourceGroup with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rgArn - The ARN of the resource group.

rgTags - The tags (key and value pairs) of the resource group. This data type property is used in the CreateResourceGroup action.

rgCreatedAt - The time at which resource group is created.

rgArn :: Lens' ResourceGroup Text

The ARN of the resource group.

rgTags :: Lens' ResourceGroup (NonEmpty ResourceGroupTag)

The tags (key and value pairs) of the resource group. This data type property is used in the CreateResourceGroup action.

rgCreatedAt :: Lens' ResourceGroup UTCTime

The time at which resource group is created.

ResourceGroupTag

data ResourceGroupTag

This data type is used as one of the elements of the ResourceGroup data type.

See: resourceGroupTag smart constructor.

resourceGroupTag

Arguments

:: Text
rgtKey

-> ResourceGroupTag

Creates a value of ResourceGroupTag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rgtValue - The value assigned to a tag key.

rgtKey - A tag key.

rgtValue :: Lens' ResourceGroupTag (Maybe Text)

The value assigned to a tag key.

rgtKey :: Lens' ResourceGroupTag Text

A tag key.

RulesPackage

data RulesPackage

Contains information about an Amazon Inspector rules package. This data type is used as the response element in the DescribeRulesPackages action.

See: rulesPackage smart constructor.

rulesPackage

Arguments

:: Text
rpArn

-> Text
rpName

-> Text
rpVersion

-> Text
rpProvider

-> RulesPackage

Creates a value of RulesPackage with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rpDescription - The description of the rules package.

rpArn - The ARN of the rules package.

rpName - The name of the rules package.

rpVersion - The version ID of the rules package.

rpProvider - The provider of the rules package.

rpDescription :: Lens' RulesPackage (Maybe Text)

The description of the rules package.

rpArn :: Lens' RulesPackage Text

The ARN of the rules package.

rpName :: Lens' RulesPackage Text

The name of the rules package.

rpVersion :: Lens' RulesPackage Text

The version ID of the rules package.

rpProvider :: Lens' RulesPackage Text

The provider of the rules package.

Subscription

data Subscription

This data type is used as a response element in the ListEventSubscriptions action.

See: subscription smart constructor.

subscription

Arguments

:: Text
sResourceARN

-> Text
sTopicARN

-> NonEmpty EventSubscription
sEventSubscriptions

-> Subscription

Creates a value of Subscription with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sResourceARN - The ARN of the assessment template that is used during the event for which the SNS notification is sent.

sTopicARN - The ARN of the Amazon Simple Notification Service (SNS) topic to which the SNS notifications are sent.

sEventSubscriptions - The list of existing event subscriptions.

sResourceARN :: Lens' Subscription Text

The ARN of the assessment template that is used during the event for which the SNS notification is sent.

sTopicARN :: Lens' Subscription Text

The ARN of the Amazon Simple Notification Service (SNS) topic to which the SNS notifications are sent.

sEventSubscriptions :: Lens' Subscription (NonEmpty EventSubscription)

The list of existing event subscriptions.

Tag

data Tag

A key and value pair. This data type is used as a request parameter in the SetTagsForResource action and a response element in the ListTagsForResource action.

See: tag smart constructor.

tag

Arguments

:: Text
tagKey

-> Tag

Creates a value of Tag with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tagValue - A value assigned to a tag key.

tagKey - A tag key.

tagValue :: Lens' Tag (Maybe Text)

A value assigned to a tag key.

tagKey :: Lens' Tag Text

A tag key.

TelemetryMetadata

data TelemetryMetadata

The metadata about the Amazon Inspector application data metrics collected by the agent. This data type is used as the response element in the GetTelemetryMetadata action.

See: telemetryMetadata smart constructor.

telemetryMetadata

Arguments

:: Text
tmMessageType

-> Integer
tmCount

-> TelemetryMetadata

Creates a value of TelemetryMetadata with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

tmDataSize - The data size of messages that the agent sends to the Amazon Inspector service.

tmMessageType - A specific type of behavioral data that is collected by the agent.

tmCount - The count of messages that the agent sends to the Amazon Inspector service.

tmDataSize :: Lens' TelemetryMetadata (Maybe Integer)

The data size of messages that the agent sends to the Amazon Inspector service.

tmMessageType :: Lens' TelemetryMetadata Text

A specific type of behavioral data that is collected by the agent.

tmCount :: Lens' TelemetryMetadata Integer

The count of messages that the agent sends to the Amazon Inspector service.

TimestampRange

data TimestampRange

This data type is used in the AssessmentRunFilter data type.

See: timestampRange smart constructor.

timestampRange :: TimestampRange

Creates a value of TimestampRange with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

trEndDate - The maximum value of the timestamp range.

trBeginDate - The minimum value of the timestamp range.

trEndDate :: Lens' TimestampRange (Maybe UTCTime)

The maximum value of the timestamp range.

trBeginDate :: Lens' TimestampRange (Maybe UTCTime)

The minimum value of the timestamp range.