Network.AWS.Inspector.UnsubscribeFromEvent

Description

Disables the process of sending Amazon Simple Notification Service (SNS) notifications about a specified event to a specified SNS topic.

Creating a Request

unsubscribeFromEvent

Arguments

:: Text
ufeResourceARN

-> InspectorEvent
ufeEvent

-> Text
ufeTopicARN

-> UnsubscribeFromEvent

Creates a value of UnsubscribeFromEvent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ufeResourceARN - The ARN of the assessment template that is used during the event for which you want to stop receiving SNS notifications.

ufeEvent - The event for which you want to stop receiving SNS notifications.

ufeTopicARN - The ARN of the SNS topic to which SNS notifications are sent.

data UnsubscribeFromEvent

See: unsubscribeFromEvent smart constructor.

Request Lenses

ufeResourceARN :: Lens' UnsubscribeFromEvent Text

The ARN of the assessment template that is used during the event for which you want to stop receiving SNS notifications.

ufeEvent :: Lens' UnsubscribeFromEvent InspectorEvent

The event for which you want to stop receiving SNS notifications.

ufeTopicARN :: Lens' UnsubscribeFromEvent Text

The ARN of the SNS topic to which SNS notifications are sent.

Destructuring the Response

unsubscribeFromEventResponse :: UnsubscribeFromEventResponse

Creates a value of UnsubscribeFromEventResponse with the minimum fields required to make a request.

data UnsubscribeFromEventResponse

See: unsubscribeFromEventResponse smart constructor.