Network.AWS.Inspector.UpdateAssessmentTarget

Description

Updates the assessment target that is specified by the ARN of the assessment target.

Creating a Request

updateAssessmentTarget

Arguments

:: Text
uatAssessmentTargetARN

-> Text
uatAssessmentTargetName

-> Text
uatResourceGroupARN

-> UpdateAssessmentTarget

Creates a value of UpdateAssessmentTarget with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uatAssessmentTargetARN - The ARN of the assessment target that you want to update.

uatAssessmentTargetName - The name of the assessment target that you want to update.

uatResourceGroupARN - The ARN of the resource group that is used to specify the new resource group to associate with the assessment target.

data UpdateAssessmentTarget

See: updateAssessmentTarget smart constructor.

Request Lenses

uatAssessmentTargetARN :: Lens' UpdateAssessmentTarget Text

The ARN of the assessment target that you want to update.

uatAssessmentTargetName :: Lens' UpdateAssessmentTarget Text

The name of the assessment target that you want to update.

uatResourceGroupARN :: Lens' UpdateAssessmentTarget Text

The ARN of the resource group that is used to specify the new resource group to associate with the assessment target.

Destructuring the Response

updateAssessmentTargetResponse :: UpdateAssessmentTargetResponse

Creates a value of UpdateAssessmentTargetResponse with the minimum fields required to make a request.

data UpdateAssessmentTargetResponse

See: updateAssessmentTargetResponse smart constructor.